% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat.R
\name{audit_hurdat}
\alias{audit_hurdat}
\title{HURDAT audit}
\usage{
audit_hurdat(df)
}
\arguments{
\item{df}{Dataframe, parsed HURDAT dataset}
}
\description{
Run an audit on the HURDAT dataset and identifies duplicate records of `Key`
  and `DateTime`.
}
\keyword{internal}
