% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat-tidy.R
\name{clean_lat}
\alias{clean_lat}
\alias{clean_lon}
\title{Clean latitude and longitude from character string to double}
\usage{
clean_lat(lat = NULL)

clean_lon(lon = NULL)
}
\arguments{
\item{lat}{character list or data.frame object containing only Latitude}

\item{lon}{character list or data.frame object containing only Longitude}
}
\value{
numeric list with positive and/or negative numbers
}
\description{
The latitude/longitude values \emph{must} contain the hemisphere identifier
     at the end otherwise bad values are introduced. In this instance, a
     new data table will be created called \code{lat_problems} or
     \code{lon_problems} and a warning message is displayed. However,
     \strong{all} values are returned.
}
\examples{
# All values except the last Latitude or Longitude are returned as
# expected. The bad value in the last row generates the lat_problems and
# lon_problems data frames.
x <- data.frame(
    ID = 1:11,
    Latitude = c("1.0S", "5S", "10.S", "28.1S", "28.15S", "1.0N", "5N",
                 "10.N", "28.1N", "28.15N", "1.0"),
    Longitude = c("1.0E", "5E", "10.E", "28.1E", "128.15E", "1.0W", "5W",
                 "10.W", "28.1W", "128.15W", "1.0"),
    stringsAsFactors = FALSE)
x$Latitude <- clean_lat(x$Latitude)
x$Longitude <- clean_lon(x$Longitude)

}

