% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CP_wind_radius}
\alias{CP_wind_radius}
\title{Central Pacific Wind Radius}
\format{A data frame of 14 variables
\describe{
     \item{\code{Key}}{Unique key identifying the tropical cyclone.
         Formatted like AABBCCCC where AA is Basin, BB is YearNum and CC is
         Year}
     \item{\code{Datetime}}{Date/time of the observation}
     \item{\code{WindRadii34NE}}{Expansion of winds > 34 kts in the
         northeastern quadrant.}
     \item{\code{WindRadii34SE}}{Expansion of winds > 34 kts in the
         southeastern quadrant.}
     \item{\code{WindRadii34SW}}{Expansion of winds > 34 kts in the
         southwestern quadrant.}
     \item{\code{WindRadii34NW}}{Expansion of winds > 34 kts in the
         northwestern quadrant.}
     \item{\code{WindRadii50NE}}{Expansion of winds > 50 kts in the
         northeastern quadrant.}
     \item{\code{WindRadii50SE}}{Expansion of winds > 50 kts in the
         southeastern quadrant.}
     \item{\code{WindRadii50SW}}{Expansion of winds > 50 kts in the
         southwestern quadrant.}
     \item{\code{WindRadii50NW}}{Expansion of winds > 50 kts in the
         northwestern quadrant.}
     \item{\code{WindRadii64NE}}{Expansion of winds > 64 kts in the
         northeastern quadrant.}
     \item{\code{WindRadii64SE}}{Expansion of winds > 64 kts in the
         southeastern quadrant.}
     \item{\code{WindRadii64SW}}{Expansion of winds > 64 kts in the
         southwestern quadrant.}
     \item{\code{WindRadii64NW}}{Expansion of winds > 64 kts in the
         northwestern quadrant.}
}

Use Key to tie this dataframe to its parent data frame. Use Key and
     Datetime to tie to sibling data frames.}
\usage{
CP_wind_radius
}
\description{
Recorded wind radius data for each observation, where available. This
     data frame is substantially smaller than its sibling "details" data
     frame because many storms simply had no data. Where there is a record
     means at least one of the wind radius variables have an integer value.
}
\keyword{datasets}

