% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone_energy.R
\name{cyclone_energy}
\alias{cyclone_energy}
\title{Accumulated Cyclone Energy}
\usage{
cyclone_energy(storm_id)
}
\arguments{
\item{storm_id}{a character string of storm ID.}
}
\description{
This function takes in a character string indicating the storm ID and returns
the accumulated cyclone energy of the selected storm.
}
\examples{
# The accumulated cyclone energy of storm AL061866
cyclone_energy("AL061866")

}
