% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EP_summary}
\alias{EP_summary}
\title{East Pacific Summary}
\format{A dataframe with nine variables:
\describe{
     \item{\code{Key}}{Unique key identifying the tropical cyclone.
         Formatted like AABBCCCC where AA is Basin, BB is YearNum and CC is
         Year}
     \item{\code{Records}}{Number of records for the storm.}
     \item{\code{StartDate}}{Date/time of first recorded observation (UTC)}
     \item{\code{EndDate}}{Date/time of last recorded observation (UTC)}
     \item{\code{YearNum}}{The number of the storm for the year}
     \item{\code{Year}}{Year of the storm}
     \item{\code{Name}}{Name of the storm (UNNAMED for early storms)}
     \item{\code{MaxWind}}{Maximum recorded wind in knots}
     \item{\code{MinPressure}}{Minimum central pressure in millibars}
}

This dataframe is the parent data frame for EP_details and EP_wind_radius.
     Key exists in all child frames to tie the data frames together.}
\usage{
EP_summary
}
\description{
Summary information on each East Pacific cyclone.
}
\keyword{datasets}
