% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HURDAT-package.R
\docType{package}
\name{HURDAT-package}
\alias{HURDAT}
\alias{HURDAT-package}
\title{HURDAT: Hurricane Re-Analysis Project}
\description{
Scraped dataset of the Hurricane Research Division's Hurricane 
    Re-Analysis Project known as HURDAT. Storm details are available for most 
    known hurricanes and tropical storms for the Atlantic and northeastern 
    Pacific ocean (northwestern hemisphere). See <http://www.aoml.noaa.gov/hrd/hurdat/Data_Storm.html> 
    for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timtrice/HURDAT}
  \item Report bugs at \url{https://github.com/timtrice/HURDAT/issues}
}

}
\author{
\strong{Maintainer}: Tim Trice \email{tim.trice@gmail.com}

Authors:
\itemize{
  \item Chris Landsea \email{Chris.Landsea@noaa.gov} [data contributor]
}

}
\keyword{internal}
