% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat.R
\docType{package}
\name{HURDAT}
\alias{HURDAT}
\alias{HURDAT-package}
\title{HURDAT}
\description{
Hurricanes is a web-scraping library for R designed to deliver
hurricane data (past and current) into well-organized datasets. With these
datasets you can explore past hurricane tracks, forecasts and structure
elements.
}
