% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfall.R
\name{is.landfall}
\alias{is.landfall}
\title{Whether Made Landfall}
\usage{
is.landfall(storm_id)
}
\arguments{
\item{storm_id}{a character string of storm ID.}
}
\value{
Integer value:
0: The storm did not make landfall.
1: The storm made landfall.
}
\description{
This function takes in a character string indicating the storm ID and returns
whether the storm made landfall in the continental United States.
}
\examples{
is.landfall("AL061851")
is.landfall("AL071851")

}
