% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat.R
\name{parse_hurdat}
\alias{parse_hurdat}
\title{parse_hurdat}
\usage{
parse_hurdat(basin)
}
\arguments{
\item{basin}{character name of basin; c("AL", "EP")}
}
\description{
Parse raw HURDAT files.
}
\keyword{internal}
