% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strom_map.R
\name{storm_map}
\alias{storm_map}
\title{Plot Map of Storms Movement Paths}
\usage{
storm_map(id)
}
\arguments{
\item{id}{a character vector of storms ID.}
}
\description{
This function takes in a character vector indicating a selection of storms ID
and returns a map recording their paths in different colors.
}
\examples{
selected.storm<-c("AL011851","AL041851")
storm_map(selected.storm)

}
