% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_map_size.R
\name{storm_map_size}
\alias{storm_map_size}
\title{Plot Map of Storm Movement Path and Size}
\usage{
storm_map_size(storm_id)
}
\arguments{
\item{storm_id}{a character string of storm ID.}
}
\description{
This function takes in a character string indicating the storm ID and returns
a map recording the movement path of the storm with different sizes and colors
indicating the size of the selected storm. The
}
\examples{
storm_map_size("AL041851")

}
