% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_track.R
\name{storm_track}
\alias{storm_track}
\title{Storm Position Track}
\usage{
storm_track(storm_id)
}
\arguments{
\item{storm_id}{a character string of storm ID.}
}
\description{
This function takes in a character string indicating the storm ID and returns
a list of the position of the storm every 30 minutes based on linear
interpolation.
}
\examples{
# Get the list of position for storm AL031869
storm_track("AL031869")

}
