% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat.R
\name{parse_hurdat}
\alias{parse_hurdat}
\title{Parse HURDAT}
\usage{
parse_hurdat(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Take a vector of lines with raw HURDAT information and convert to a
  dataframe object.
}
\keyword{internal}
