% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat-parse.R
\name{raw_hurdat}
\alias{raw_hurdat}
\title{Retrieve Raw HURDAT files for Atlantic or the  Northeast Pacific and North
     central Pacific basins}
\usage{
raw_hurdat(basins = list(atlantic = c(TRUE, al_hurdat2()), nepac = c(TRUE,
  ep_hurdat2())))
}
\arguments{
\item{basins}{a multidimensional list of basins each containing a boolean
on whether or not to load the data (TRUE or FALSE) and a url to the
dataset.}
}
\value{
Returns a data frame for each basin, slightly restructured. An
     effort is made to keep the data in it's original format as much as
     possible. However, each storm has a header row. This header row is the
     only row that identifies the storm info (i.e., Name). Without it,
     there is no way to differentiate storms. Therefore, the header row is
     split and left-binded to the following rows. Therefore, if you perform
     a QA-check on nrow of the create dataframe versus nrow of the original
     text files, the created data frames will be shorter but have more
     columns. Otherwise, data will still need to be cleaned.
}
\description{
Raw text files \emph{should} be found at
     \url{http://www.nhc.noaa.gov/data/hurdat/} as of this writing. The
     codebook for Atlantic HURDAT can be found at
     \url{http://www.nhc.noaa.gov/data/hurdat/hurdat2-format-atlantic.pdf};
     East and Central Pacific at
     \url{http://www.nhc.noaa.gov/data/hurdat/hurdat2-format-nencpac.pdf}.
}
\details{
HURDAT files are updated typically over the winter or early spring. As of
     this writing, only the Atlantic HURDAT file has been updated.

By default, calling \code{raw_hurdat()} will retrieve two HURDAT datasets;
     one for the Atlantic basin and one for both the north central and
     north east Pacific basin. See Examples.

You can choose to load only Atlantic or the north east Pacific and north
     central \emph{and} north east Pacific basin datasets. You can subset
     the combined Pacific dataset by \code{Basin == CP}, discarding the
     rest.

Problem warnings are to be expected. In the raw text files, the header
     only contain four comma-delimited values where otherwise there are
     21. \code{readr} will generate a warning of n parsing failures and
     state that it expected 21 colums but only received four. \emph{This
     warning is the only acceptable warning}. When any \code{problems()}
     are generated for a dataset, a data frame will be created showing
     these problems. Provided that all rows reflect the message above,
     there is no issue.
}
\examples{

# Load only Atlantic basin data
datasets <- list("atlantic" = c(TRUE, al_hurdat2()))
raw_hurdat(datasets)

# Load only Pacific basin data
datasets <- list("nepac" = c(TRUE, ep_hurdat2()))
raw_hurdat(datasets)

# Load only Atlantic and northeast Pacific basin data
datasets <- list("atlantic" = c(TRUE, al_hurdat2()),
                 "nepac" = c(TRUE, ep_hurdat2()))
raw_hurdat(datasets)

# Default
datasets <- list("atlantic" = c(TRUE, al_hurdat2()),
                 "nepac" = c(TRUE, ep_hurdat2()))
raw_hurdat(datasets)

}
\seealso{
Atlantic codebook: \url{http://www.nhc.noaa.gov/data/hurdat/hurdat2-format-atlantic.pdf}

NE/NC Pacific codebook: \url{http://www.nhc.noaa.gov/data/hurdat/hurdat2-format-atlantic.pdf}
}
